fix: refresh yielded lifecycle generated surfaces
